submissions-team/aditi-phadnis/Frameworks/requirements.txt
openai
python-dotenv
streamlit
langchain 
openai
langchain-community
langchain-openai
pygame
SpeechRecognition
pydub